requirements.txt
# FastAPI + Server
fastapi
uvicorn[standard]
pydantic==2.7.4
python-multipart==0.0.9

# LangGraph + LangChain (compatible versions)
langgraph==0.0.62
langchain-core==0.2.27
langchain-openai==0.1.7

# LLM + OpenAI SDK
openai==1.40.1

# Utilities
python-dotenv==1.0.1
jinja2==3.1.4
pylatexenc==2.10
httpx==0.27.0

langsmith

